Correctly update Store on dynamic add cens/conts
The store is upadated based on the instructions produced by the
leviathan_cen:lm_compare/2. Those instructions has to be aligned
to the leviathan_store data model - this happens in
leviathan_store:filter_and_sort_update_instructions/1. Then, the
filtered instrcutions, are translated to the leviathan_db updtes,
that are finally invoked in a single transaciton.